Revamped API calls
