+++ lib/textrazor.rb
@@ -4,6 +4,7 @@ require "textrazor/util"
 require "textrazor/dictionary"
 require "textrazor/client"
 require "textrazor/request"
+require "textrazor/api_response"
 require "textrazor/response"
 require "textrazor/category"
 require "textrazor/topic"
